Move elements into separate file
