+++ src/lib.rs
@@ -30,8 +30,11 @@
 
 use std::collections::HashMap;
 
-use egui::{self, epaint, text::LayoutJob, Id, NumExt, Pos2, RichText, Sense, TextStyle, Ui, Vec2};
+use egui::{self, text::LayoutJob, Id, Pos2, RichText, TextStyle, Ui, Vec2};
 use pulldown_cmark::{CowStr, HeadingLevel, Options};
+
+mod elements;
+use elements::*;
 
 #[cfg(feature = "better_syntax_highlighting")]
 use syntect::{
@@ -999,70 +1002,7 @@ impl CommonMarkViewerInternal {
                     ui.fonts(|f| f.layout_job(job))
                 };
 
-                let pre_text_edit_position = ui.next_widget_position();
-                let mut text = block.content.strip_suffix('\n').unwrap_or(&block.content);
-
-                // To manually add background color to the code block, we imitate what
-                // TextEdit does internally
-                let where_to_put_background = ui.painter().add(egui::Shape::Noop);
-
-                // We use a `TextEdit` to make the text selectable.
-                // Note that we take a `&mut` to a non-`mut` `&str`, which is
-                // the how to tell `egui` that the text is not editable.
-                let output = egui::TextEdit::multiline(&mut text)
-                    .layouter(&mut layout)
-                    .desired_width(max_width)
-                    // prevent trailing lines
-                    .desired_rows(1)
-                    .show(ui);
-
-                // Background color + frame (This is lost when TextEdit it not editable)
-                let frame_rect = output.response.rect;
-                ui.painter().set(
-                    where_to_put_background,
-                    epaint::RectShape::new(
-                        frame_rect,
-                        ui.style().noninteractive().rounding,
-                        ui.visuals().extreme_bg_color,
-                        ui.visuals().widgets.noninteractive.bg_stroke,
-                    ),
-                );
-
-                // Copy icon
-                let spacing = &ui.style().spacing;
-                let position = pre_text_edit_position
-                    + egui::vec2(
-                        max_width - spacing.icon_width - spacing.button_padding.x,
-                        -spacing.icon_width + spacing.button_padding.y,
-                    );
-
-                if ui
-                    .put(
-                        egui::Rect {
-                            min: position,
-                            max: position,
-                        },
-                        egui::Button::new("🗐")
-                            .small()
-                            .frame(false)
-                            .fill(egui::Color32::TRANSPARENT),
-                    )
-                    .clicked()
-                {
-                    use egui::TextBuffer as _;
-                    let copy_text = if let Some(cursor) = output.cursor_range {
-                        let selected_chars = cursor.as_sorted_char_range();
-                        let selected_text = text.char_range(selected_chars);
-                        if selected_text.is_empty() {
-                            text.to_owned()
-                        } else {
-                            selected_text.to_owned()
-                        }
-                    } else {
-                        text.to_owned()
-                    };
-                    ui.ctx().copy_text(copy_text);
-                }
+                code_block(ui, max_width, &block.content, &mut layout);
             });
         }
 
@@ -1162,133 +1102,11 @@ fn syntect_color_to_egui(color: syntect::highlighting::Color) -> egui::Color32 {
     egui::Color32::from_rgb(color.r, color.g, color.b)
 }
 
-fn newline(ui: &mut Ui) {
-    ui.label("\n");
-}
-
-fn bullet_point(ui: &mut Ui) {
-    let (rect, _) = ui.allocate_exact_size(
-        egui::vec2(width_body_space(ui) * 4.0, height_body(ui)),
-        Sense::hover(),
-    );
-    ui.painter().circle_filled(
-        rect.center(),
-        rect.height() / 6.0,
-        ui.visuals().strong_text_color(),
-    );
-}
-
-fn bullet_point_hollow(ui: &mut Ui) {
-    let (rect, _) = ui.allocate_exact_size(
-        egui::vec2(width_body_space(ui) * 4.0, height_body(ui)),
-        Sense::hover(),
-    );
-    ui.painter().circle(
-        rect.center(),
-        rect.height() / 6.0,
-        egui::Color32::TRANSPARENT,
-        egui::Stroke::new(0.6, ui.visuals().strong_text_color()),
-    );
-}
-
-fn number_point(ui: &mut Ui, number: &str) {
-    let (rect, _) = ui.allocate_exact_size(
-        egui::vec2(width_body_space(ui) * 4.0, height_body(ui)),
-        Sense::hover(),
-    );
-    ui.painter().text(
-        rect.right_center(),
-        egui::Align2::RIGHT_CENTER,
-        format!("{number}. "),
-        TextStyle::Body.resolve(ui.style()),
-        ui.visuals().strong_text_color(),
-    );
-}
-
-fn footnote_start(ui: &mut Ui, note: &str) {
-    ui.label(RichText::new(note).raised().strong().small());
-}
-
-fn footnote(ui: &mut Ui, text: &str) {
-    let (rect, _) = ui.allocate_exact_size(
-        egui::vec2(width_body_space(ui) * 4.0, height_body(ui)),
-        Sense::hover(),
-    );
-    ui.painter().text(
-        rect.right_top(),
-        egui::Align2::RIGHT_TOP,
-        format!("{text}."),
-        TextStyle::Small.resolve(ui.style()),
-        ui.visuals().strong_text_color(),
-    );
-}
-
-fn height_body(ui: &Ui) -> f32 {
-    ui.text_style_height(&TextStyle::Body)
-}
-
-fn width_body_space(ui: &Ui) -> f32 {
-    let id = TextStyle::Body.resolve(ui.style());
-    ui.fonts(|f| f.glyph_width(&id, ' '))
-}
-
 #[cfg(feature = "better_syntax_highlighting")]
 fn default_theme(ui: &Ui) -> &str {
     if ui.style().visuals.dark_mode {
         DEFAULT_THEME_DARK
     } else {
         DEFAULT_THEME_LIGHT
-    }
-}
-
-// Stripped down version of egui's Checkbox. The only difference is that this
-// creates a noninteractive checkbox. ui.add_enabled could have been used instead,
-// but it makes the checkbox too grey.
-struct Checkbox<'a> {
-    checked: &'a mut bool,
-}
-
-impl<'a> Checkbox<'a> {
-    pub fn without_text(checked: &'a mut bool) -> Self {
-        Checkbox { checked }
-    }
-}
-
-impl<'a> egui::Widget for Checkbox<'a> {
-    fn ui(self, ui: &mut Ui) -> egui::Response {
-        let Checkbox { checked } = self;
-
-        let spacing = &ui.spacing();
-        let icon_width = spacing.icon_width;
-
-        let mut desired_size = egui::vec2(icon_width, 0.0);
-        desired_size = desired_size.at_least(Vec2::splat(spacing.interact_size.y));
-        desired_size.y = desired_size.y.max(icon_width);
-        let (rect, response) = ui.allocate_exact_size(desired_size, Sense::click());
-
-        if ui.is_rect_visible(rect) {
-            let visuals = ui.style().visuals.noninteractive();
-            let (small_icon_rect, big_icon_rect) = ui.spacing().icon_rectangles(rect);
-            ui.painter().add(epaint::RectShape::new(
-                big_icon_rect.expand(visuals.expansion),
-                visuals.rounding,
-                visuals.bg_fill,
-                visuals.bg_stroke,
-            ));
-
-            if *checked {
-                // Check mark:
-                ui.painter().add(egui::Shape::line(
-                    vec![
-                        egui::pos2(small_icon_rect.left(), small_icon_rect.center().y),
-                        egui::pos2(small_icon_rect.center().x, small_icon_rect.bottom()),
-                        egui::pos2(small_icon_rect.right(), small_icon_rect.top()),
-                    ],
-                    visuals.fg_stroke,
-                ));
-            }
-        }
-
-        response
     }
 }
